The base class for all steps

Override this class to create a new step.

TicStep

Format

An object of class R6ClassGenerator of length 24.

Methods

run

This method must be overridden, it is called when running the stage to which a step has been added.

prepare

This method is called when preparing the stage to which a step has been added. Override this method to install any R packages your step might need, because this allows them to be cached for subsequential runs.

check

This method determines if a step is prepared and run. Return FALSE if conditions for running this step are not met.