Process a tic.R file

Return a named list of stages generated by source()ing a file that uses the tic DSL. The file is sourced only once per session, unless its modification timestamp changes.

load_from_file(path = "tic.R", ..., mtime = file.mtime(path))

Arguments

path
[string]
Path to the stage definition file, default: "tic.R".

...
ignored

mtime
[POSIXt]
For internal use, to detect changes to this file.