Step: Build a bookdown book

Build a bookdown book using bookdown::render_book().

step_build_bookdown(...)

Arguments

...
Arguments passed on to bookdown::render_book

input

An input filename (or multiple filenames). If preview = TRUE, only files specified in this argument are rendered, otherwise all R Markdown files specified by the book are rendered.

output_format

Arguments to be passed to rmarkdown::render(). For preview_chapter(), ... is passed to render_book().

clean

Arguments to be passed to rmarkdown::render(). For preview_chapter(), ... is passed to render_book().

envir

Arguments to be passed to rmarkdown::render(). For preview_chapter(), ... is passed to render_book().

clean_envir

Whether to clean up the environment envir before rendering the book. By default, the environment is cleaned when rendering the book in a non-interactive R session.

output_dir

The output directory. If NULL, a field named output_dir in the configuration file _bookdown.yml will be used (possibly not specified, either, in which case a directory name _book will be used).

new_session

Whether to use new R sessions to compile individual Rmd files (if not provided, the value of the new_session option in _bookdown.yml is used; if this is also not provided, new_session = FALSE).

preview

Whether to render and preview the input files specified by the input argument. Previewing a certain chapter may save compilation time as you actively work on this chapter, but the output may not be accurate (e.g. cross-references to other chapters will not work).

encoding

Ignored. The character encoding of all input files is supposed to be UTF-8.

config_file

The book configuration file.

Examples

dsl_init()

#> ✔ Creating a blank tic stage configuration

get_stage("script") %>%
  add_step(step_build_bookdown("."))

dsl_get()

#> ── script ───────────────────────────────────────────────────────────── stage ──
#> ▶ step_build_bookdown(".")