Step: Build pkgdown documentation

Builds package documentation with the pkgdown package.

step_build_pkgdown(...)

Arguments

...
Arguments passed on to pkgdown::build_site

pkg

Path to package.

examples

Run examples?

document

If TRUE, will run devtools::document() before updating the site.

run_dont_run

Run examples that are surrounded in \dontrun?

seed

Seed used to initialize so that random examples are reproducible.

lazy

If TRUE, will only rebuild articles and reference pages if the source is newer than the destination.

override

An optional named list used to temporarily override values in _pkgdown.yml

preview

If TRUE, or is.na(preview) && interactive(), will preview freshly generated section in browser.

new_process

If TRUE, will run build_site() in a separate process. This enhances reproducibility by ensuring nothing that you have loaded in the current process affects the build process.

See also