Step: Install an SSH key

Writes a private SSH key encoded in an environment variable to a file in ~/.ssh. Only run in non-interactive settings and if the environment variable exists and is non-empty. The use_travis_deploy() and use_tic() functions encode a private key as an environment variable for use with this function.

step_install_ssh_keys(name = "id_rsa")

Arguments

name
[string]
Name of the environment variable and the target file, default: "id_rsa".

See also

Examples

dsl_init()

#> ✔ Creating a blank tic stage configuration

get_stage("before_deploy") %>%
  add_step(step_install_ssh_keys())

dsl_get()

#> ── before_deploy ────────────────────────────────────────────────────── stage ──
#> ▶ step_install_ssh_keys()