Step: Run arbitrary R code

Captures the expression and executes it when running the step. An optional preparatory expression can be provided that is executed during preparation. If the top-level expression is a qualified function call (of the format package::fun()), the package is installed during preparation.

step_run_code(call = NULL, prepare_call = NULL)

Arguments

call
[call]
An arbitrary R expression executed during the stage to which this step is added. The default is useful if you only pass prepare_call.

prepare_call
[call]
An optional arbitrary R expression executed during preparation.

See also

Examples

step_run_code(update.packages(ask = FALSE))

#> <RunCode>
#>   Inherits from: <TicStep>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     initialize: function (call, prepare_call = NULL) 
#>     prepare: function () 
#>     run: function () 
#>   Private:
#>     call: call
#>     install_call_dep: function (call) 
#>     prepare_call: NULL
#>     seed: 123

# Will install covr from CRAN during preparation:
step_run_code(covr::codecov())

#> <RunCode>
#>   Inherits from: <TicStep>
#>   Public:
#>     check: function () 
#>     clone: function (deep = FALSE) 
#>     initialize: function (call, prepare_call = NULL) 
#>     prepare: function () 
#>     run: function () 
#>   Private:
#>     call: call
#>     install_call_dep: function (call) 
#>     prepare_call: NULL
#>     seed: 123