tic: Tasks Integrating Continuously: CI-Agnostic Workflow Definitions

Provides a way to describe common build and deployment workflows for R-based projects: packages, websites (e.g. blogdown, pkgdown), or data processing (e.g. research compendia). The recipe is described independent of the continuous integration tool used for processing the workflow (e.g. 'Travis CI' or 'AppVeyor').

Details

The use_tic() function prepares a code repository for use with this package. See DSL for an overview of tic's domain-specific language for defining stages and steps, step_hello_world() and the links therein for available steps, and macro for an overview over the available macros that bundle several steps.

See also