Use CI YAML templates

Installs YAML templates for various CI providers.

use_travis_yml(type)

use_appveyor_yml(type)

use_circle_yml(type)

Arguments

type
[character]
Which template to use. The string should be given following the logic <platform>-<action>. See details for more.

Type

tic supports a variety of different YAML templates which follow the <platform>-<action> pattern. The first one is mandatory, the others are optional.

Possible values for <provider> are travis, and circle

Possible values for <platform> are linux, and macos, windows.

Possible values for <action> are matrix and deploy.

Not every combinations is supported on all CI systems. For example, for use_appveyor_yaml() only windows and windows-matrix are valid.

Here is a list of all available combinations:

Provider	Operating system	Deployment	multiple R versions	Call

----------	------------------	------------	---------------------	---------------------------------------------------------

Travis	Linux	no	no	use_travis_yml("linux")

Linux	yes	no	use_travis_yml("linux-deploy")

Linux	no	yes	use_travis_yml("linux-matrix")

Linux	yes	yes	use_travis_yml("linux-deploy-matrix")

macOS	no	no	use_travis_yml("macOS")

macOS	yes	no	use_travis_yml("macos-deploy")

macOS	no	yes	use_travis_yml("macos-matrix")

macOS	yes	yes	use_travis_yml("macos-deploy-matrix")

----------	------------------	------------	---------------------	---------------------------------------------------------

Circle	Linux	no	no	use_circle_yml("linux")

Linux	yes	no	use_travis_yml("linux-deploy")

Linux	no	yes	use_travis_yml("linux-matrix")

Linux	no	yes	use_travis_yml("linux-deploy-matrix")

----------	------------------	------------	---------------------	---------------------------------------------------------

Appveyor	Windows	no	no	use_appveyor_yml("windows")

Windows	no	yes	use_travis_yml("windows-matrix")