tic's domain-specific language

Functions to define stages and their constitutent steps.

get_stage() returns a Stage object for a stage given by name. This function only works when called by load_from_file().

add_step() adds a step to a stage, see step_hello_world() and the links therein for available steps.

add_code_step() is a shortcut for add_step(step_run_code(...)).

add_package_checks() adds default steps related to package checks to the "before_install", "install", "script" and "after_success" stages:

A step_rcmdcheck() in the "script" stage, using the warnings_are_errors, notes_are_errors and args arguments.

A call to covr::codecov() in the "after_success" stage (only for non-interactive CIs)

get_stage(name)

add_step(stage, step)

add_code_step(stage, call, prepare_call = NULL)

add_package_checks(warnings_are_errors = TRUE,
  notes_are_errors = FALSE, args = c("--no-manual", "--as_cran"))

Arguments

name
[string]
The name for the stage.

stage
[Stage]
A Stage object as returned by get_stage().

step
[function]
A function that constructs a Step object, such as step_hello_world().

call
[call]
An arbitrary expression executed during the stage to which this step is added.

prepare_call
[call]
An optional arbitrary expression executed during preparation.

warnings_are_errors
[flag]
Should warnings be treated as errors? Default: TRUE.

notes_are_errors
[flag]
Should notes be treated as errors? Default: FALSE.

args
[character]
Passed to [rcmdcheck::rcmdcheck()] (after splitting at spaces), default: "--no-manual --as-cran".