Render a R Markdown README and deploy to Github

do_readme_rmd() renders an R Markdown README and deploys the rendered README.md file to Github. It adds default steps to the "before_deploy" and "deploy" stages:

step_setup_ssh() in the "before_deploy" to setup the upcoming deployment

step_setup_push_deploy() in the "before_deploy" stage

rmarkdown::render() in the "deploy" stage

step_do_push_deploy() in the "deploy" stage.

Usage

do_readme_rmd(
  checkout = TRUE,
  remote_url = NULL,
  commit_message = NULL,
  force = FALSE,
  private_key_name = "TIC_DEPLOY_KEY"
)

Arguments

checkout

[flag]
Check out the current contents of the repository? Defaults to TRUE, set to FALSE if the build process relies on existing contents or if you deploy to a different branch.

remote_url

[string]
The URL of the remote Git repository to push to, defaults to the current GitHub repository.

commit_message

[string]
Commit message to use, defaults to a useful message linking to the CI build and avoiding recursive CI runs.

force

[logical]
Add --force flag to git commands?

private_key_name

string
Only needed when deploying from builds on GitHub Actions. If you have set a custom name for the private key during creation of the SSH key pair via tic::use_ghactions_deploy()] or use_tic(), pass this name here.

See also

Examples

if (FALSE) {
dsl_init()

do_readme_rmd()

dsl_get()
}