Storage for stages and steps

Stages and steps defined using tic's DSL are stored in an internal object in the package. The stages are accessible through dsl_get(). When running the stages, by default a configuration defined in the tic.R file is loaded with dsl_load(). See use_tic() for setting up a tic.R file.

For interactive tests, an empty storage can be initialized with dsl_init(). This happens automatically the first time dsl_get() is called (directly or indirectly).

dsl_get()

dsl_load(path = "tic.R", force = FALSE, quiet = FALSE)

dsl_init(quiet = FALSE)

Arguments

path
[string]
Path to the stage definition file, default: "tic.R".

force
[flag]
Set to TRUE to force loading from file even if a configuration exists. By default an existing configuration is not overwritten by dsl_load().

quiet
[flag]
Set to TRUE to turn off verbose output.

Value

A named list of opaque stage objects with a "class" attribute and a corresponding print() method for pretty output. Use the high-level get_stage() and add_step() functions to configure, and the stages functions to run.

Examples

dsl_init()

#> ✔ Creating a blank tic stage configuration

dsl_get()

#> ℹ No steps defined in any stage

dsl_load(system.file("templates/package/tic.R", package = "tic"))

#> ℹ Using existing tic stage configuration, use `force = TRUE` to reload

dsl_load(system.file("templates/package/tic.R", package = "tic"), force = TRUE)

#> ✔ Loading tic stage configuration from /home/travis/build/ropenscilabs/tic/inst/templates/package/tic.R

dsl_get()

#> ── install ──────────────────────────────────────────────────────────── stage ──
#> ▶ step_install_deps(repos = repo_default())
#> ▶ step_install_deps(repos = repo_default())
#> ── script ───────────────────────────────────────────────────────────── stage ──
#> ▶ step_rcmdcheck(warnings_are_errors = NULL, notes_are_errors = NULL, 
#>     args = NULL, build_args = NULL, error_on = "warning", repos = repo_default(), 
#>     timeout = Inf)
#> ── after_success ────────────────────────────────────────────────────── stage ──
#> ▶ step_run_code(covr::codecov(quiet = FALSE))
#> ── before_deploy ────────────────────────────────────────────────────── stage ──
#> ▶ step_setup_ssh()
#> ▶ step_setup_push_deploy(path = "docs", branch = "gh-pages", remote_url = NULL, 
#>     orphan = FALSE, checkout = TRUE)
#> ── deploy ───────────────────────────────────────────────────────────── stage ──
#> ▶ step_build_pkgdown()
#> ▶ step_do_push_deploy(path = "docs", commit_message = NULL, commit_paths = ".")