Macros

The DSL offers a fine-grained interface to the individual stages of a CI run. Macros are tic's way of adding several related steps to the relevant stages. All macros use the do_ prefix.

The do_package_checks() macro adds default checks for R packages, including installation of dependencies and running a test coverage analysis.

The do_pkgdown() macro adds the necessary steps for building and deploying pkgdown documentation for a package.

The do_bookdown() macro adds the necessary steps for building and deploying a bookdown book.