Step: Install packages

These steps are useful if your CI run needs packages which are not declared as dependencies in your DESCRIPTION. Usually you should declare these dependencies, but this may not always be desired.

A step_install_cran() step installs one package from CRAN via install.packages(), but only if it's not already installed.

A step_install_github() step installs one or more packages from GitHub via remotes::install_github(), the packages are only installed if their GitHub version is different from the locally installed version.

step_install_cran(package = NULL, ...)

step_install_github(repo = NULL, ...)

Arguments

package
Package(s) to install

...
Passed on to install.packages() or remotes::install_github().

repo
Package to install in the "user/repo" format.

See also