Step: Install packages

These steps are useful if your CI run needs additional packages. Usually they are declared as dependencies in your DESCRIPTION, but it is also possible to install dependencies manually. By default, binary versions of packages are installed if possible, even if the CRAN version is ahead.

A step_install_deps() step installs all package dependencies declared in DESCRIPTION, using remotes::install_deps(). This includes upgrading outdated packages.

A step_install_cran() step installs one package from CRAN via install.packages(), but only if it's not already installed.

A step_install_github() step installs one or more packages from GitHub via remotes::install_github(), the packages are only installed if their GitHub version is different from the locally installed version.

step_install_deps(repos = repo_default(), type = NULL)

step_install_cran(package = NULL, ..., repos = repo_default(), type = NULL)

step_install_github(repo = NULL, ..., type = NULL)

Arguments

repos
CRAN-like repositories to install from, defaults to repo_default().

type
Passed on to install.packages(). The default avoids installation from source on Windows and macOS by passing .Platform$pkgType.

package
Package(s) to install

...
Passed on to install.packages() or remotes::install_github().

repo
Package to install in the "user/repo" format.

See also

Examples

dsl_init()

#> ✔ Creating a blank tic stage configuration

get_stage("install") %>%
  add_step(step_install_deps())

dsl_get()

#> ── install ──────────────────────────────────────────────────────────── stage ──
#> ▶ step_install_deps()

dsl_init()

#> ✔ Creating a blank tic stage configuration

get_stage("install") %>%
  add_step(step_install_cran("magick"))

dsl_get()

#> ── install ──────────────────────────────────────────────────────────── stage ──
#> ▶ step_install_cran("magick")

dsl_init()

#> ✔ Creating a blank tic stage configuration

get_stage("install") %>%
  add_step(step_install_github("rstudio/gt"))

dsl_get()

#> ── install ──────────────────────────────────────────────────────────── stage ──
#> ▶ step_install_github("rstudio/gt")