Step: Check a package

Check a package using rcmdcheck, which ultimately calls R CMD check. The preparation consists of installing package dependencies via remotes::install_deps() with dependencies = TRUE, and updating all packages.

step_rcmdcheck(warnings_are_errors = TRUE, notes_are_errors = FALSE,
  args = "--no-manual --as-cran")

Arguments

warnings_are_errors
[flag]
Should warnings be treated as errors? Default: TRUE.

notes_are_errors
[flag]
Should notes be treated as errors? Default: FALSE.

args
[character]
Passed to [rcmdcheck::rcmdcheck()] (after splitting at spaces), default: "--no-manual --as-cran".

Details

This step uses a dedicated library, a subdirectory tic-pkg of the current user library (the first element of .libPaths()), for the checks. This is done to minimize conflicts between dependent packages and packages that are required for running the various steps.