Step: Setup push deploy

Clones a repo, inits author information, and sets up remotes for a subsequent step_do_push_deploy().

step_setup_push_deploy(path = ".", branch = ci()$get_branch(),
  orphan = FALSE, remote_url = paste0("git@github.com:", ci()$get_slug(),
  ".git"), checkout = TRUE)

Arguments

path
[string] Path to the repository, default "." which means setting up the current repository.

branch
[string] Target branch, default: current branch

orphan
[flag] Create and force-push an orphan branch consisting of only one commit? This can be useful e.g. for path = "docs", branch = "gh-pages", but cannot be applied for pushing to the current branch.

remote_url
[string] The URL of the remote Git repository to push to, defaults to the current GitHub repository.

checkout
[flag] Check out the current contents of the repository? Defaults to TRUE, set to FALSE if the build process relies on existing contents or if you deploy to a different branch.

See also

Other deploy steps: step_do_push_deploy, step_push_deploy

Other steps: step_add_to_drat, step_add_to_known_hosts, step_build_pkgdown, step_do_push_deploy, step_hello_world, step_install_ssh_keys, step_push_deploy, step_run_code, step_setup_ssh, step_test_ssh, step_write_text_file