Initialize CI testing using tic

Prepares a repo for building and deploying supported by tic.

use_tic(
  wizard = interactive(),
  linux = "travis",
  mac = "travis",
  windows = "appveyor",
  deploy = "travis",
  matrix = "none",
  private_key_name = "TIC_DEPLOY_KEY",
  travis_endpoint = ".com",
  quiet = FALSE
)

Arguments

wizard
[flag]
Interactive operation? If TRUE, a menu will be shown.

linux
[string]
Which CI provider(s) to use to test on Linux. Possible options are "travis", "circle", "ghactions", "none" and "all".

mac
[string]
Which CI provider(s) to use to test on macOS Possible options are "none", "travis" and "ghactions".

windows
[string]
Which CI provider(s) to use to test on Windows Possible options are "none", "appveyor" and "ghactions".

deploy
[string]
Which CI provider(s) to use to deploy artifacts such as pkgdown documentation. Possible options are "travis", "circle", "ghactions", "none" and "all".

matrix
[string]
For which CI provider(s) to set up matrix builds. Possible options are "travis", "circle", "ghactions", "none" and "all".

private_key_name
string
Only needed when deploying from builds on Travis CI or GitHub Actions. If you have set a custom name for the private key during creation of the SSH key pair via travis::use_travis_deploy(), tic::use_ghactions_deploy() or use_tic(), pass this name here.

travis_endpoint
[string]
The Travis CI endpoint to use. Possible options are ".org" and ".com". Default is ".com". See travis::travis_enable() for more information.

quiet
[flag]
Less verbose output? Default: FALSE.

Details

Query information which CI providers should be used

Setup permissions for providers selected for deployment

Create YAML files for selected providers

Create a default tic.R file depending on the repo type (package, website, bookdown, ...)

Examples

# Requires interactive mode
if (FALSE) {
  use_tic()

  # Pre-specified settings favoring Circle CI:
  use_tic(
    wizard = FALSE,
    linux = "circle",
    mac = "travis",
    windows = "appveyor",
    deploy = "circle",
    matrix = "all"
  )
}