Use CI YAML templates

Installs YAML templates for various CI providers. These function are also used within use_tic().

use_travis_yml(type = "linux-macos-deploy-matrix")

use_appveyor_yml(type = "windows")

use_circle_yml(type = "linux-matrix-deploy")

use_ghactions_yml(type = "linux-macos-windows-deploy")

Arguments

type
[character]
Which template to use. The string should be given following the logic <platform>-<action>. See details for more.

pkgdown

If type contains "deploy", tic by default also sets the environment variable BUILD_PKGDOWN=true. This triggers a call to pkgdown::build_site() via the do_pkgdown macro in tic.R for the respective runners.

If a setting includes "matrix" and builds on multiple R versions, the job building on R release is chosen to build the pkgdown site.

YAML Type

tic supports a variety of different YAML templates which follow the <platform>-<action> pattern. The first one is mandatory, the others are optional.

Possible values for <platform> are linux, and macos, windows.

Possible values for <action> are matrix and deploy.

Not every combinations is supported on all CI systems. For example, for use_appveyor_yaml() only windows and windows-matrix are valid.

For backward compatibility use_ghactions_yml() will be default build and deploy on all platforms.

Here is a list of all available combinations:

Provider	Operating system	Deployment	multiple R versions	Call

----------	------------------	------------	---------------------	---------------------------------------------------------

Travis	Linux	no	no	use_travis_yml("linux")

Linux	yes	no	use_travis_yml("linux-deploy")

Linux	no	yes	use_travis_yml("linux-matrix")

Linux	yes	yes	use_travis_yml("linux-deploy-matrix")

macOS	no	no	use_travis_yml("macos")

macOS	yes	no	use_travis_yml("macos-deploy")

macOS	no	yes	use_travis_yml("macos-matrix")

macOS	yes	yes	use_travis_yml("macos-deploy-matrix")

Linux + macOS	no	no	use_travis_yml("linux-macos")

Linux + macOS	yes	no	use_travis_yml("linux-macos-deploy")

Linux + macOS	no	yes	use_travis_yml("linux-macos-matrix")

Linux + macOS	yes	yes	use_travis_yml("linux-macos-deploy-matrix")

----------	------------------	------------	---------------------	---------------------------------------------------------

Circle	Linux	no	no	use_circle_yml("linux")

Linux	yes	no	use_travis_yml("linux-deploy")

Linux	no	yes	use_travis_yml("linux-matrix")

Linux	no	yes	use_travis_yml("linux-deploy-matrix")

----------	------------------	------------	---------------------	---------------------------------------------------------

Appveyor	Windows	no	no	use_appveyor_yml("windows")

Windows	no	yes	use_travis_yml("windows-matrix")

----------	------------------	------------	---------------------	---------------------------------------------------------

GH Actions	Linux	no	no	use_ghactions_yml("linux") -

Linux	yes	no	use_ghactions_yml("linux-deploy)

macOS	no	no	use_ghactions_yml("macos)

macOS	yes	no	use_ghactions_yml("macos-deploy)

Windows	no	no	use_ghactions_yml("windows)

Windows	yes	no	use_ghactions_yml("windows-deploy)

Linux + macOS	no	no	use_ghactions_yml("linux-macos")

Linux + macOS	yes	no	use_ghactions_yml("linux-macos-deploy")

Linux + Windows	no	no	use_ghactions_yml("linux-windows")

Linux + Windows	yes	no	use_ghactions_yml("linux-windows-deploy")

macOS + Windows	no	no	use_ghactions_yml("macos-windows")

macOS + Windows	yes	no	use_ghactions_yml("macos-windows-deploy")

Linux + macOS + Windows	no	no	use_ghactions_yml("linux-macos-windows")

Linux + macOS + Windows	yes	no	use_ghactions_yml("linux-macos-windows-deploy")